Look Back in Angular
New projects for ancient platforms which now can be improved so as not to suck so bad. My Meteoric rise. Quasar kills the radio star. Vue router gets satisfactory results and  so now on to themes and stuff. Advancing on the flexbox front which reminds me of Peircean semiotics. A scheme to memorize.