Back to the Vueture
Routers for Vuetify and Vue Material and painful efforts to grok the latter. In search of time not necessarily lost on rudimentary projects hoovered in their tens of gigas from a previous Arch LInux box and actually not all of them all that bad. How to make dough and earn bread on the Word-Wild Web. I ponder this.