Add files via upload
The Meteors of Whittier and why Daemonite and BMD are right for me and maybe too Semantic for Vue. And some frustrating adventures with Nuxt, resolved for the moment setting an alternative port to serve n. Collecting workable PWAs and SPAs and SSRs. Pruning the hard drive after a zero disk-space alarm. No need to panic.