My Semantic Configurathon
Too many screenshots did not view.  But I want to remember just how it was I configured Semantic UI for Vue and Webpack and Meteor and others in a recent hacking session. Hacking as in a hack golfer. Whack it with idiotic brute force until it finally reaches the green.  Then putter about some more.